requirements.txt
google-genai>=1.0.0
lxml>=5.0.0
cssselect>=1.2.0
requests>=2.31.0
streamlit>=1.40.0
